somewhat closer to fixed typing
