+++ lighthouse-treemap/types/treemap.d.ts
@@ -1,35 +1,12 @@
+import '../../types/externs';
 import _TreemapUtil = require('../app/src/util.js');
-import '../../types/lhr';
-import '../../types/audit-details';
 
 declare global {
-  module Treemap {
-    interface Options {
-      lhr: LH.Result;
-    }
-
-    interface RootNodeContainer {
-      /** Arbitrary name identifier. Usually a script url. */
-      name: string;
-      node: Node;
-    }
-
-    interface Node {
-      /** Arbitrary name identifier. Usually a path component from a source map. */
-      name: string;
-      resourceBytes: number;
-      unusedBytes?: number;
-      /** If present, this module is a duplicate. String is normalized source path. See ModuleDuplication.normalizeSource */
-      duplicatedNormalizedModuleName?: string;
-      children?: Node[];
-    }
-  }
-
   interface WebTreeMapOptions {
     padding: [number, number, number, number];
     spacing: number;
-    caption(node: Treemap.Node): string;
-    showNode?(node: Treemap.Node): boolean;
+    caption(node: LH.Treemap.Node): string;
+    showNode?(node: LH.Treemap.Node): boolean;
   }
 
   var webtreemap: {
@@ -40,7 +17,7 @@ declare global {
   var TreemapUtil: typeof _TreemapUtil;
 
   interface Window {
-    __treemapOptions?: Treemap.Options;
+    __treemapOptions?: LH.Treemap.Options;
   }
 }
